# Better perf for many (empty) components