## Description of the pull request:
This is the result of me trying to figure out whether or not it makes sense to pursue #1376 and #1379 lightweight components any further. I've come to the conclusion that it's not really needed because the introduction of proper models in 0.8 basically killed the model-related overhead involved with components. They're still not super-lightweight because they're roughly 20x more expensive than a plain partial, but to be fair, partial performance is __way__ better than it was before 0.8. The only bits I could think to strip off of components are somewhat necessary to have an encapsulated widget anyways (its own potentially isolated model, registries, and lifecycle events).

When I started profiling with an iterated list of 1000 components with just the letter 'c' as their template and partials with just the letter 'p' as their content, the initial render for the components was ~240ms settling to ~140ms, and the initial render for the partials was ~27ms settling to ~4ms. That's a pretty wide gap. With these changes, the initial render for components goes to ~150ms settling to ~50ms, and the partial times stay about the same.

The biggest change comes from a slight refactor on the code that checks to see if elements can be detached at the end of a runloop, namely, if there are no nodes to detach, there's no need to try to gather outros to check against an empty list.

The second comes in the form of a slight event firing refactor wherein instances that have no namespaced events are entirely skipped during bubbling, so you don't pay a perf penalty unless you're actually using namespaced events. This only accounts for 5-7% of the gap.

The last bit comes from fixing some de-opts around `combine` and `extend` and re-jiggering a few hot codepaths to avoid extra function calls and whatnot.

My final test with 50000 iterations of each shows that for huge numbers of iterations, components are ~10x more expensive than partials on initial run. edge actually bails when trying with 50000, but 20000 took 23523ms where this took 1188ms. 50000 components on this takes ~2800ms. 20000 partials takes ~200ms on edge and ~160ms on this for the initial run.

Now, I'd imagine that components that _actually do something_ other than render a single character are probably even _more_ expensive than partials, but I was trying to get a feel for just how much overhead there is at the minimum implementation. I feel pretty good about this PR though, given the difference from edge when rendering 20000 components.

## Fixes the following issues:
To some extent, any pertaining to performance with lots of components.

## Is breaking:
Nope.