@evs-chris 2016-12-30
It looks like the lightweight components proposal was made somewhere around 0.6.0. Roughly duplicating  the test described in #1376, it looks like this PR using components is ~1.7x inline and partials in 0.6.0. 0.7.3 had a decent performance improvement all around, except when removing a bunch of components, which was particularly bad for some reason (like 3x render time to unrender). The 0.8 series improved general rendering further over 0.7.3, but it only improved component rendering by about 10%. Given that, I'm going to go ahead and merge this and say that we're now to the point that at 3x inline time, plain old components are probably fast enough not to need a special slightly hobbled version of them, at least for performance purposes.